requirements.txt
openai>=1.10.0
yt-dlp>=2023.11.16
python-dotenv>=1.0.1
tqdm>=4.66.1
easyocr>=1.7.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
streamlit>=1.28.0

